git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18867 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/xwindow.c
@@ -1151,7 +1151,7 @@ MagickPrivate XVisualInfo *XBestVisualInfo(Display *display,
 #define XVisualColormapSize(visual_info) MagickMin((unsigned int) (\
   (visual_info->klass == TrueColor) || (visual_info->klass == DirectColor) ? \
    visual_info->red_mask | visual_info->green_mask | visual_info->blue_mask : \
-   (unsigned int) visual_info->colormap_size),1U << visual_info->depth)
+   (unsigned long) visual_info->colormap_size),1UL << visual_info->depth)
 
   char
     *map_type,
